Update partitioning to support tryboot to auto-revert
This updates the `fwup` configuration to support the Raspberry Pi
tryboot feature so that it's possible to recover from messed up Linux
kernels and OTP boot failures. The way this works is that after the
firmware update is applied, the system is rebooted one time into the new
firmware slot. If the firmware slot is good, then the app needs to call
`Nerves.Runtime.validate_firmware` to make it so that subsequent boots
use the new slot. The theory is that the app won't validate the firmware
unless it's good, so that if anything crashes, the previously working
firmware will be used.

Even though the partitioning looks different, it's possible to upgrade
non-tryboot Nerves Raspberry Pi firmware without losing application
data. This is supported by lining up partitions so that they're close to
the same places. There's no going back, though. Also, there's a window
where cancelling the firmware update can brick the Pi due to the rootfs
partition needing to be shifted, so be careful.